requirements.txt
Babel==2.17.0
PyQt5-sip==12.12.1
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3
pyserial==3.5
